minion/requirements.txt
#### requirements.txt ####
#
#### Requirements for boomerang ####
flask
pyminizip
Ghost.py
requests
verbalexpressions
#
